test. step5 - basicGame과 컨트롤러 분리로 테스트 변경
